feat: l1/l2 deploy commands for testnet/mainnet
