+++ config/contexts/v0.0.9.yaml
@@ -1,0 +1,135 @@
+# Devnet context to be used for local deployments against Anvil chain
+version: 0.0.9
+context:
+  # Name of the context
+  name: "devnet"
+  # Chains available to this context
+  chains:
+    l1:
+      chain_id: 31337
+      rpc_url: "http://localhost:8545"
+      fork:
+        block: 8825097
+        url: ""
+        block_time: 3
+    l2:
+      chain_id: 31338
+      rpc_url: "http://localhost:9545"
+      fork:
+        block: 28753690
+        url: ""
+        block_time: 3
+  # Stake Root Transporter configuration
+  transporter:
+    schedule: "0 */2 * * *"
+    private_key: "0x5f8e6420b9cb0c940e3d3f8b99177980785906d16fb3571f70d7a05ecf5f2172"
+    bls_private_key: "0x5f8e6420b9cb0c940e3d3f8b99177980785906d16fb3571f70d7a05ecf5f2172"
+    active_stake_roots: []
+  # All key material (BLS and ECDSA) within this file should be used for local testing ONLY
+  # ECDSA keys used are from Anvil's private key set
+  # BLS keystores are deterministically pre-generated and embedded. These are NOT derived from a secure seed
+  # Available private keys for deploying
+  deployer_private_key: "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80" # Anvil Private Key 0
+  app_private_key: "0x5de4111afa1a4b94908f83103eb1f1706367c2e68ca870fc3fb9a804cdab365a" # Anvil Private Key 2
+  # List of stakers and their delegations 
+  stakers:
+    - address: "0x23618e81E3f5cdF7f54C3d65f7FBc0aBf5B21E8f"
+      ecdsa_key: "0xdbda1821b80551c9d65939329250298aa3472ba22feea921c0cf5d620ea67b97" # Anvil 8
+      deposits:
+        - strategy_address: "0x8b29d91e67b013e855EaFe0ad704aC4Ab086a574"
+          name: "stETH_Strategy"
+          deposit_amount: "5ETH" # depositIntoStrategy amount 
+      operator: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65" # Operator to delegate the stake via delegationManager.delegateTo()
+    - address: "0xa0Ee7A142d267C1f36714E4a8F75612F20a79720"
+      ecdsa_key: "0x2a871d0798f97d79848a013d4936a73bf4cc922c825d33c1cf7073dff6d409c6"
+      deposits:
+        - strategy_address: "0x8b29d91e67b013e855EaFe0ad704aC4Ab086a574"
+          name: "stETH_Strategy"
+          deposit_amount: "5ETH" # depositIntoStrategy amount 
+      operator: "0x90F79bf6EB2c4f870365E785982E1f101E93b906"
+  # List of Operators and their private keys / stake details
+  operators:
+    - address: "0x90F79bf6EB2c4f870365E785982E1f101E93b906"
+      ecdsa_key: "0x7c852118294e51e653712a81e05800f419141751be58f605c371e15141b007a6" # Anvil Private Key 3
+      ecdsa_keystore_path: "keystores/operator1.ecdsa.keystore.json"
+      ecdsa_keystore_password: "testpass"
+      bls_keystore_path: "keystores/operator1.bls.keystore.json"
+      bls_keystore_password: "testpass"
+      allocations:
+        - strategy_address: "0x8b29d91e67b013e855EaFe0ad704aC4Ab086a574"
+          name: "stETH_Strategy"
+          # Only allocate if these operator set IDs exist in the deployed operator_sets 
+          operator_set_allocations:
+            - operator_set: "0"
+              allocation_in_wads: "500000000000000000" # 5e17 i.e 50% of max allocation
+            - operator_set: "1"
+              allocation_in_wads: "500000000000000000" # 5e17 i.e 50% of max allocation
+    - address: "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65"
+      ecdsa_key: "0x47e179ec197488593b187f80a00eb0da91f1b9d0b13f8733639f19c30a34926a" # Anvil Private Key 4
+      ecdsa_keystore_path: "keystores/operator2.ecdsa.keystore.json"
+      ecdsa_keystore_password: "testpass"
+      bls_keystore_path: "keystores/operator2.bls.keystore.json"
+      bls_keystore_password: "testpass"
+      allocations:
+        - strategy_address: "0x8b29d91e67b013e855EaFe0ad704aC4Ab086a574"
+          name: "stETH_Strategy"
+          # Only allocate if these operator set IDs exist in the deployed operator_sets 
+          operator_set_allocations:
+            - operator_set: "0" 
+              allocation_in_wads: "500000000000000000" # 5e17 i.e 50% of max allocation
+            - operator_set: "1"
+              allocation_in_wads: "500000000000000000" # 5e17 i.e 50% of max allocation
+    - address: "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc"
+      ecdsa_key: "0x8b3a350cf5c34c9194ca85829a2df0ec3153be0318b5e2d3348e872092edffba" # Anvil Private Key 5
+      ecdsa_keystore_path: "keystores/operator3.ecdsa.keystore.json"
+      ecdsa_keystore_password: "testpass"
+      bls_keystore_path: "keystores/operator3.bls.keystore.json"
+      bls_keystore_password: "testpass"
+    - address: "0x976EA74026E726554dB657fA54763abd0C3a0aa9"
+      ecdsa_key: "0x92db14e403b83dfe3df233f83dfa3a0d7096f21ca9b0d6d6b8d88b2b4ec1564e" # Anvil Private Key 6
+      ecdsa_keystore_path: "keystores/operator4.ecdsa.keystore.json"
+      ecdsa_keystore_password: "testpass"
+      bls_keystore_path: "keystores/operator4.bls.keystore.json"
+      bls_keystore_password: "testpass"
+    - address: "0x14dC79964da2C08b23698B3D3cc7Ca32193d9955"
+      ecdsa_key: "0x4bbbf85ce3377467afe5d46f804f221813b2bb87f24d81f60f1fcdbf7cbf4356" # Anvil Private Key 7
+      ecdsa_keystore_path: "keystores/operator5.ecdsa.keystore.json"
+      ecdsa_keystore_password: "testpass"
+      bls_keystore_path: "keystores/operator5.bls.keystore.json"
+      bls_keystore_password: "testpass"
+  # AVS configuration
+  avs:
+    address: "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"
+    avs_private_key: "0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d" # Anvil Private Key 1
+    metadata_url: "https://my-org.com/avs/metadata.json"
+    registrar_address: "0x0123456789abcdef0123456789ABCDEF01234567"
+  # Core EigenLayer contract addresses
+  eigenlayer:
+    l1:
+      allocation_manager: "0x42583067658071247ec8CE0A516A58f682002d07"
+      delegation_manager: "0xD4A7E1Bd8015057293f0D0A557088c286942e84b"
+      strategy_manager: "0x2E3D6c0744b10eb0A4e6F679F71554a39Ec47a5D"
+      bn254_table_calculator: "0xc2c0bc13571aC5115709C332dc7AE666606b08E8"
+      cross_chain_registry: "0x287381B1570d9048c4B4C7EC94d21dDb8Aa1352a"
+      key_registrar: "0xA4dB30D08d8bbcA00D40600bee9F029984dB162a"
+      release_manager: "0xd9Cb89F1993292dEC2F973934bC63B0f2A702776"
+      operator_table_updater: "0xB02A15c6Bd0882b35e9936A9579f35FB26E11476"
+    l2:
+      bn254_certificate_verifier: "0xff58A373c18268F483C1F5cA03Cf885c0C43373a"
+      operator_table_updater: "0xB02A15c6Bd0882b35e9936A9579f35FB26E11476"
+      ecdsa_certificate_verifier: "0xb3Cd1A457dEa9A9A6F6406c6419B1c326670A96F"
+  # L1 Contracts deployed on `devnet start`
+  deployed_l1_contracts: []
+  # L2 Contracts deployed on `devnet start`
+  deployed_l2_contracts: []
+  # Operator Sets registered on `devnet start`
+  operator_sets: []
+  # Operators registered on `devnet start`
+  operator_registrations: []
+  # Release artifact
+  artifact:
+    artifactId: ""
+    component: ""
+    digest: ""
+    registry: ""
+    version: ""
