+++ pkg/common/config_test.go
@@ -83,8 +83,8 @@ func TestLoadConfigWithContextConfig_FromCopiedTempFile(t *testing.T) {
 	assert.Equal(t, "http://localhost:8545", cfg.Context["devnet"].Chains["l1"].RPCURL)
 	assert.Equal(t, "http://localhost:9545", cfg.Context["devnet"].Chains["l2"].RPCURL)
 
-	assert.Equal(t, 8713384, cfg.Context["devnet"].Chains["l1"].Fork.Block)
-	assert.Equal(t, 28069764, cfg.Context["devnet"].Chains["l2"].Fork.Block)
+	assert.Equal(t, 8825097, cfg.Context["devnet"].Chains["l1"].Fork.Block)
+	assert.Equal(t, 28753690, cfg.Context["devnet"].Chains["l2"].Fork.Block)
 
 	assert.Equal(t, "0x70997970C51812dc3A010C7d01b50e0d17dc79C8", cfg.Context["devnet"].Avs.Address)
 	assert.Equal(t, "0x0123456789abcdef0123456789ABCDEF01234567", cfg.Context["devnet"].Avs.RegistrarAddress)
