+++ pkg/common/devnet/constants.go
@@ -7,8 +7,6 @@ const L2_CHAIN_ARGS = "--gas-limit 140000000 --base-fee 0 --gas-price 1000000"
 
 const FUND_VALUE = "1000000000000000000"
 const DEVNET_CONTEXT = "devnet"
-const L1 = "l1"
-const L2 = "l2"
 const ANVIL_1_KEY = "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80"
 const ANVIL_2_KEY = "0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d"
