+++ pkg/common/devnet/getters.go
@@ -137,76 +137,6 @@ func GetDevnetRPCUrlDefault(cfg *common.ConfigWithContextConfig, chainName strin
 	return chainConfig.RPCURL, nil
 }
 
-func GetDevnetForkUrlDefault(cfg *common.ConfigWithContextConfig, chainName string) (string, error) {
-	// Check in env first for L1 fork url
-	l1ForkUrl := os.Getenv("L1_FORK_URL")
-	if chainName == "l1" && l1ForkUrl != "" {
-		return l1ForkUrl, nil
-	}
-
-	// Check in env first for l2 fork url
-	l2ForkUrl := os.Getenv("L2_FORK_URL")
-	if chainName == "l2" && l2ForkUrl != "" {
-		return l2ForkUrl, nil
-	}
-
-	// Fallback to context defined value
-	chainConfig, found := cfg.Context[DEVNET_CONTEXT].Chains[chainName]
-	if !found {
-		return "", fmt.Errorf("failed to get chainConfig for chainName : %s", chainName)
-	}
-	if chainConfig.Fork.Url == "" {
-		return "", fmt.Errorf("fork-url not set for %s; set fork-url in ./config/context/devnet.yaml or .env and consult README for guidance", chainName)
-	}
-	return chainConfig.Fork.Url, nil
-}
-
-// GetEigenLayerAddresses returns EigenLayer L1 addresses from the context config
-// Falls back to constants if not found in context
-func GetEigenLayerAddresses(cfg *common.ConfigWithContextConfig) (allocationManager, delegationManager string, strategyManager string, keyRegistrar string, crossChainRegistry string, bn254TableCalculator string, releaseManager string) {
-	if cfg == nil || cfg.Context == nil {
-		return ALLOCATION_MANAGER_ADDRESS, DELEGATION_MANAGER_ADDRESS, STRATEGY_MANAGER_ADDRESS, KEY_REGISTRAR_ADDRESS, CROSS_CHAIN_REGISTRY_ADDRESS, BN254_TABLE_CALCULATOR_ADDRESS, RELEASE_MANAGER_ADDRESS
-	}
-
-	devnetCtx, found := cfg.Context[DEVNET_CONTEXT]
-	if !found || devnetCtx.EigenLayer == nil {
-		return ALLOCATION_MANAGER_ADDRESS, DELEGATION_MANAGER_ADDRESS, STRATEGY_MANAGER_ADDRESS, KEY_REGISTRAR_ADDRESS, CROSS_CHAIN_REGISTRY_ADDRESS, BN254_TABLE_CALCULATOR_ADDRESS, RELEASE_MANAGER_ADDRESS
-	}
-
-	allocationManager = devnetCtx.EigenLayer.L1.AllocationManager
-	if allocationManager == "" {
-		allocationManager = ALLOCATION_MANAGER_ADDRESS
-	}
-
-	delegationManager = devnetCtx.EigenLayer.L1.DelegationManager
-	if delegationManager == "" {
-		delegationManager = DELEGATION_MANAGER_ADDRESS
-	}
-	strategyManager = devnetCtx.EigenLayer.L1.StrategyManager
-	if strategyManager == "" {
-		strategyManager = STRATEGY_MANAGER_ADDRESS
-	}
-	keyRegistrar = devnetCtx.EigenLayer.L1.KeyRegistrar
-	if keyRegistrar == "" {
-		keyRegistrar = KEY_REGISTRAR_ADDRESS
-	}
-
-	crossChainRegistry = devnetCtx.EigenLayer.L1.CrossChainRegistry
-	if crossChainRegistry == "" {
-		crossChainRegistry = CROSS_CHAIN_REGISTRY_ADDRESS
-	}
-	bn254TableCalculator = devnetCtx.EigenLayer.L1.BN254TableCalculator
-	if bn254TableCalculator == "" {
-		bn254TableCalculator = BN254_TABLE_CALCULATOR_ADDRESS
-	}
-	releaseManager = devnetCtx.EigenLayer.L1.ReleaseManager
-	if releaseManager == "" {
-		releaseManager = RELEASE_MANAGER_ADDRESS
-	}
-
-	return allocationManager, delegationManager, strategyManager, keyRegistrar, crossChainRegistry, bn254TableCalculator, releaseManager
-}
-
 // GetL1Port returns the L1 devnet port (default port + 0)
 func GetL1Port(basePort int) int {
 	return basePort
